Fix compiler failure in Xcode 11
For some unknown reason (at least to me) the `subscript(position:)`
extension on `MutableCollection` where `Index == Int` was causing the
compilation to fail with a weird `DESERIALIZATION FAILURE` error,
as mentioned on #15.

Updating the `subscript` to use `C.Index` instead of `Int` appears to
make the compiler happy and successfully compile the project. My
suspicion is that it somehow enables it to disambiguate between the
multiple `subscript` implementations, but it's pure speculation 🔮.

On our tests this seems to have fixed the issue, and since it's an
equivalent definition of the subscript (even though more "precise"
from a generics POV), it should cause no side effects.

Fixes #15. 🎉