Make `NonEmpty` conform to `Collection` in type declaration
Apparently by defining `NonEmpty` as conforming to `Collection` right
away in its declaration (instead of in an extension), the compilation
failure seems to be fixed.

Even when performing a clean build, the target now successfully
compiles. 🎉

Not sure about the actual cause for this, but this definitely seems
like a compiler bug (file ordering / dependency, perhaps?).